issue 674 Optimize code logic
